requirements.txt
astroid==2.12.13
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
dill==0.3.6
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
pylint==2.15.6
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.6
urllib3==1.26.13
wrapt==1.14.1
